extract `server.ts`
